Разбит на модули data редусер
