+++ src/components/app/app.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import {connect} from 'react-redux';
-import {ActionCreator} from '../../reducer/data/data.js';
-import {getAllOffers, getIsOffersLoading, getCurrentCity, getOffers, getCities} from '../../reducer/data/selectors';
+import {ActionCreator} from '../../reducer/data/action-creator.js';
+import {getAllOffers, getIsOffersLoading, getCurrentCity, getOffers, getCitiesState} from '../../reducer/data/selectors';
 import {getIsAuthorizationRequired} from '../../reducer/user/selectors';
 import {PageScreen} from '../page-screen/page-screen.jsx';
 
@@ -18,7 +18,7 @@ const mapStateToProps = (state, ownProps) => Object.assign({}, ownProps, {
   isOffersLoading: getIsOffersLoading(state),
   currentCity: getCurrentCity(state),
   offers: getOffers(state),
-  cities: getCities(state),
+  cities: getCitiesState(state),
   isAuthorizationRequired: getIsAuthorizationRequired(state),
 });
