+++ src/reducer/data/action-type.js
@@ -1,0 +1,7 @@
+export const ActionType = {
+  LOAD_ALL_OFFERS: `LOAD_ALL_OFFERS`,
+  CHANGE_LOADING_STATE: `CHANGE_LOADING_STATE`,
+  CHANGE_CITY: `CHANGE_CITY`,
+  GET_OFFERS_LIST: `GET_OFFERS_LIST`,
+  GET_CITIES_LIST: `GET_CITIES_LIST`,
+};
