+++ src/reducer/data/data.test.js
@@ -1,6 +1,8 @@
 import MockAdapter from 'axios-mock-adapter';
 import createAPI from '../../api.js';
-import {ActionType, ActionCreator, Operation, prepareData} from "./data.js";
+import {ActionType} from './action-type.js';
+import {ActionCreator} from './action-creator.js';
+import {Operation, prepareData} from "./data.js";
 import {offers} from '../../mocks/offers.js';
 
 describe(`Reducer works correctly`, () => {
